requirements.txt
streamlit
pandas
plotly
folium
scikit-learn
numpy
altairstreamlit-folium>=0.15.0
